requirements.txt
coverage==4.5.4
dj-database-url==0.5.0
Django==2.2.6
djangorestframework==3.10.3
gunicorn==20.0.0
psycopg2==2.8.4
python-decouple==3.1
pytz==2019.3
reverse==0.1.0
sqlparse==0.3.0
whatever==0.5
whitenoise==4.1.4
